Turn the emoji picker into a web component
